requirements.txt
pandas
sqlalchemy
dask
prefect
psycopg2-binary
django
djangorestframework
django-cors-headers